+++ src/main/kotlin/ui/setup/screens/GameDirectoryScreen.kt
@@ -55,3 +55,4 @@ fun GameDirectoryScreen(routerState: RouterState, config: MutableState<SetupStat
             }
         }
     }
+